Move read_memory() tests to test_common.py
