+++ test_wasm.py
@@ -6,28 +6,6 @@ from wasmtime import Store, Memory, MemoryType, Limits
 
 
 class TestStreamdalWasm:
-    def test_read_memory_within_bounds(self):
-        """Test reading within bounds of memory"""
-        store = Store()
-        memory = Memory(store, MemoryType(Limits(1, 1024)))
-        data = b"Hello, World!\xa6\xa6\xa6"
-        memory.write(store, data, 0)
-
-        result = common.read_memory(memory, store, 0, len(data))
-        assert result == b"Hello, World!"
-
-    def test_read_memory_with_interspersed_pointers(self):
-        """Test reading with null pointers"""
-        store = Store()
-        memory = Memory(store, MemoryType(Limits(1, 1)))
-        data = b"Hel\xa6lo,\xa6 W\xa6orld!\xa6\xa6\xa6"
-        memory.write(store, data, 0)
-
-        result = common.read_memory(memory, store, 0)
-        assert (
-            result == b"Hel\xa6lo,\xa6 W\xa6orld!"
-        )  # Should NOT stop at the third terminator character
-
     def test_call_wasm_failure(self, mocker):
         mocker.patch(
             "streamdal.StreamdalClient._exec_wasm",
